# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/rounds": {
            "get": {
                "description": "Get round snapshots, optionally filtered by ID or season_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Get round snapshots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Snapshot ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Season ID",
                        "name": "season_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RoundSnapshot"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Take a snapshot of current racer standings for a season round",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Take round snapshot",
                "parameters": [
                    {
                        "description": "Round snapshot data (race_name, season_id, round)",
                        "name": "snapshot",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a round snapshot by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Delete round snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snapshot ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Update the scores for a draft round. Only positions \u003c 900 are counted as finished.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Update round scores (draft only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Round snapshot ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Array of updated scores",
                        "name": "scores",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RoundSnapshotScore"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/rounds/batch": {
            "get": {
                "description": "Get multiple round snapshots with scores by IDs (comma-separated)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Get round snapshots batch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated snapshot IDs",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RoundSnapshot"
                            }
                        }
                    }
                }
            }
        },
        "/api/rounds/finalize": {
            "patch": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Lock a draft round and update driver totals (points, stats, spins).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Finalize a round",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Round snapshot ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/seasons/archive": {
            "post": {
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "description": "Archive a season by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Seasons"
                ],
                "summary": "Archive season",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Season ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.RoundSnapshot": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "race_date": {
                    "type": "string"
                },
                "race_name": {
                    "type": "string"
                },
                "round": {
                    "type": "integer"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RoundSnapshotScore"
                    }
                },
                "season_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.RoundSnapshotScore": {
            "type": "object",
            "properties": {
                "dnf": {
                    "type": "boolean"
                },
                "dns": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "overheated": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "racer_id": {
                    "type": "integer"
                },
                "racer_name": {
                    "type": "string"
                },
                "snapshot_id": {
                    "type": "integer"
                },
                "spins": {
                    "type": "integer"
                }
            }
        },
...
    }
}